builder/requirements-builder.txt
invokeai-python
aiohttp
gitpython
requests
pyyaml